Setup our own translator without override the existing one
Instead of fully override the translator set by the user (or another
Sphinx extension), we create a new class dynamically by inheriting the
translator that's defined and adding our own Mixin class with our
custom behavior.

This way, we are compatible with other extensions nicely.

As a example, I tested this with
pydata-sphinx-theme (https://github.com/pandas-dev/pydata-sphinx-theme)
that defines a custom translator.

Besides, this new approach is compatible with people using HTML5
translator by default since we are not forcing HTML4 when inheriting
our own custom class anymore.