Translator on current builder
We set our own translator in the translators defined by other extensions, but
also for the current builder.

It may happen that there is an extension that defines two translators for
`dirhml` and `json`, and we are going to override them --the problem comes when
`sphinx-build` is called with `-b custombuilder` since we weren't overriding it
in that case.

This commit override the ones defined by the user, but also the default one.