Ignore non-HTML builders when resolving references
It does not make sense to override the domain classes when using other non-HTML
builders because there is nothing to add.

Although, due to a initialization race condition, we are not able to "not
override the Domains if the builder is non-HTML" so we need to skip our custom
resolution during build time.