Override `missing-reference` from intersphinx to support hoverxref
Disconnect original `missing-reference` event from `sphinx.ext.intersphinx` to
connect our own function. When this event is triggered, we call the original
`missing_reference` function and add some extra data (`data-url` and CSS
classes) to the node returned by the original function.

This data is later used by the transformer to render the node into HTML as we expect.