Use Sphinx 3.4.x to build the docs
Sphinx 3.5.x includes a feature to only include the JS and CSS on the pages that
they are used. This conflicts when we render content that uses MathJax, since
the page that shows the tooltip does not has MathJax loaded, but the content
rendered inside the tooltip requires it to work.

See https://github.com/sphinx-doc/sphinx/pull/8631