Do not remove `reftitle` because it breaks other extensions
https://github.com/readthedocs/sphinx-hoverxref/pull/86#issuecomment-852136211

There are other extensions that expect `reftitle` to be an attribute in the
node. They are doing `node['reftitle']` without checking if it exists first.

This happens in `sphinx.ext.inheritance_diagram` which is a built-in extension.
So it's preferrable to not modify the node and find another way to remove the
browser's tooltip later.

See https://github.com/sphinx-doc/sphinx/blob/4.x/sphinx/ext/inheritance_diagram.py#L413