Intersphinx: improve robustness
We were generating the inventory's key manually based on the `reftype` and
fallbacks. However, we weren't covering all the possible cases. This commit,
uses `.objtypes_for_role` domain's method to get all the possible names for a
particular `reftype` on a domain.

With this change, it also adds supports for generic `obj` type in a domain,
making our code to find all the possible alternative keys in that particular
domain. This is useful when using `default_role = 'obj'` as requested in
https://github.com/readthedocs/sphinx-hoverxref/issues/170