Remove `team-reviewers` because it's failing
```
Error: Reviews may only be requested from collaborators. One or more of the
teams you specified is not a collaborator of the readthedocs/sphinx-hoverxref
repository.
```

It seems it's a permission issue. See https://github.com/peter-evans/create-pull-request/issues/155#issuecomment-611904487

However, we are not having this issue in .org. I'm removing the reviewers for
now just to move forward because we haven't updated the requirements in a while now.