Don't disconnect html-page-context event from sphinx-tabs (#318)
We are disconnecting the html-page-context event from sphinx-tabs,
which was used to delete the static files from html files
that didn't use sphinx-tabs. But in https://github.com/executablebooks/sphinx-tabs/pull/198
this was changed to instead decide if the page should have the static files or not.

Since we are removing that event, sphinx-tabs never gets to inject the
static files, this breaks sphinx-tabs for projects using hoverxref.

We no longer need to disconnect that signal, since we are overriding the html assets policy to `always`,
which sphnx-tabs respects already https://github.com/executablebooks/sphinx-tabs/pull/132
(more than 3 years ago).